string/wstring->LPTSTR conversion is now switched by UNICODE macro.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/trunk/OpenRTM-aist@1660 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/coil/common/stringutil.cpp
@@ -416,7 +416,7 @@ namespace coil
   {
     if (ignore_case) { toLower(value); }
 
-    for (int i(0), len(list.size()); i < len; ++i)
+    for (int i(0), len(static_cast<int>(list.size())); i < len; ++i)
       {
         std::string str(list[i]);
         if (ignore_case) { toLower(str); }
